requirements.txt
# Core dependencies
torch>=1.9.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Configuration and data handling
pyyaml>=5.4.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Imbalanced learning
imbalanced-learn>=0.8.0

# Optional: For advanced visualization
# umap-learn>=0.5.0  # Uncomment if using UMAP for dimensionality reduction

# Development tools (optional)
# jupyter>=1.0.0
# notebook>=6.4.0
# ipykernel>=6.0.0

